flatten setup and nix config
