+++ src/providers/go.rs
@@ -2,7 +2,7 @@ use super::Provider;
 use crate::nixpacks::{
     app::App,
     environment::Environment,
-    nix::{NixConfig, Pkg},
+    nix::{Pkg},
     phase::{InstallPhase, SetupPhase, StartPhase},
 };
 use anyhow::Result;
@@ -19,10 +19,10 @@ impl Provider for GolangProvider {
     }
 
     fn setup(&self, _app: &App, _env: &Environment) -> Result<Option<SetupPhase>> {
-        Ok(Some(SetupPhase::new(NixConfig::new(vec![
+        Ok(Some(SetupPhase::new(vec![
             Pkg::new("pkgs.stdenv"),
             Pkg::new("pkgs.go"),
-        ]))))
+        ])))
     }
 
     fn install(&self, app: &App, _env: &Environment) -> Result<Option<InstallPhase>> {
