+++ src/providers/npm.rs
@@ -4,7 +4,7 @@ use super::Provider;
 use crate::nixpacks::{
     app::App,
     environment::{Environment, EnvironmentVariables},
-    nix::{NixConfig, Pkg},
+    nix::{Pkg},
     phase::{BuildPhase, InstallPhase, SetupPhase, StartPhase},
 };
 use anyhow::{bail, Result};
@@ -38,10 +38,10 @@ impl Provider for NpmProvider {
         let package_json: PackageJson = app.read_json("package.json")?;
         let node_pkg = NpmProvider::get_nix_node_pkg(&package_json)?;
 
-        Ok(Some(SetupPhase::new(NixConfig::new(vec![
+        Ok(Some(SetupPhase::new(vec![
             Pkg::new("pkgs.stdenv"),
             node_pkg,
-        ]))))
+        ])))
     }
 
     fn install(&self, app: &App, _env: &Environment) -> Result<Option<InstallPhase>> {
