+++ src/providers/python.rs
@@ -8,7 +8,6 @@ use crate::{
     nixpacks::{
         app::App,
         environment::Environment,
-        nix::NixConfig,
         phase::{InstallPhase, SetupPhase, StartPhase},
     },
     Pkg,
@@ -33,9 +32,7 @@ impl Provider for PythonProvider {
         _app: &App,
         _env: &crate::nixpacks::environment::Environment,
     ) -> Result<Option<SetupPhase>> {
-        Ok(Some(SetupPhase::new(NixConfig::new(vec![Pkg::new(
-            "pkgs.python38",
-        )]))))
+        Ok(Some(SetupPhase::new(vec![Pkg::new("pkgs.python38")])))
     }
 
     fn install(&self, app: &App, _env: &Environment) -> Result<Option<InstallPhase>> {
